pyproject.toml
[tool.poetry]
name = "boss-battles"
version = "0.1.0"
description = ""
authors = ["Mr. Gallo <daniel.gallo@ycdsbk12.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
